+++ application/cypress.config.ts
@@ -28,6 +28,7 @@ export default defineConfig({
           LOGIN_PASSWORD: process.env.CYPRESS_LOGIN_PASSWORD,
           PROJECT_KEY: process.env.CYPRESS_PROJECT_KEY,
           PACKAGE_NAME: process.env.CYPRESS_PACKAGE_NAME,
+          LOCALE: process.env.CYPRESS_LOCALE || 'en-GB',
         },
       };
     },