+++ application/src/components/method-details/method-details.tsx
@@ -1,8 +1,10 @@
 import { useIntl } from 'react-intl';
-import { useParams } from 'react-router-dom';
+import { Route, Switch, useParams, useRouteMatch } from 'react-router-dom';
 import {
   PageNotFound,
   CustomFormModalPage,
+  TabularModalPage,
+  TabHeader,
 } from '@commercetools-frontend/application-components';
 import { ContentNotification } from '@commercetools-uikit/notifications';
 import Text from '@commercetools-uikit/text';
@@ -20,10 +22,14 @@ import { TMethodObjectValueFormValues } from '../../types';
 import { useShowNotification } from '@commercetools-frontend/actions-global';
 import {
   DOMAINS,
+  NO_VALUE_FALLBACK,
   NOTIFICATION_KINDS_SIDE,
 } from '@commercetools-frontend/constants';
 import SelectField from '@commercetools-uikit/select-field';
 import { ApplicationPageTitle } from '@commercetools-frontend/application-shell';
+import { useIsAuthorized } from '@commercetools-frontend/permissions';
+import { PERMISSIONS } from '../../constants';
+import { formatLocalizedString } from '@commercetools-frontend/l10n';
 
 type TMethodDetailsProps = {
   onClose: () => void;
@@ -31,15 +37,18 @@ type TMethodDetailsProps = {
 
 const MethodDetails = (props: TMethodDetailsProps) => {
   const intl = useIntl();
+  const match = useRouteMatch();
   const params = useParams<{ id: string }>();
   const { loading, error, method } = useCustomObjectDetailsFetcher(params.id);
-  const { dataLocale } = useApplicationContext((context) => ({
+  const { dataLocale, projectLanguages } = useApplicationContext((context) => ({
     dataLocale: context.dataLocale ?? '',
     projectLanguages: context.project?.languages ?? [],
   }));
   const customObjectUpdater = useCustomObjectDetailsUpdater();
   const showNotification = useShowNotification();
-  const canManage = true;
+  const canManage = useIsAuthorized({
+    demandedPermissions: [PERMISSIONS.Manage],
+  });
 
   const handleSubmit = async (formikValues: TMethodObjectValueFormValues) => {
     try {
@@ -53,7 +62,17 @@ const MethodDetails = (props: TMethodDetailsProps) => {
           kind: 'success',
           domain: DOMAINS.SIDE,
           text: intl.formatMessage(messages.methodDetailsUpdated, {
-            methodName: formikValues.description,
+            methodName: formatLocalizedString(
+              {
+                name: formikValues.name,
+              },
+              {
+                key: 'name',
+                locale: dataLocale,
+                fallbackOrder: projectLanguages,
+                fallback: NO_VALUE_FALLBACK,
+              }
+            ),
           }),
         });
       }
@@ -68,22 +87,27 @@ const MethodDetails = (props: TMethodDetailsProps) => {
   ) => {
     try {
       if (method?.container && method?.key && formikValues) {
+        let clonedValues = { ...formikValues, ...{ status: status } };
         await customObjectUpdater.execute({
           container: method?.container,
           key: method?.key,
-          value: JSON.stringify({
-            id: formikValues.id,
-            description: formikValues.description,
-            status: status,
-            imageUrl: formikValues.imageUrl,
-            displayOrder: formikValues.displayOrder,
-          }),
+          value: JSON.stringify(clonedValues),
         });
         showNotification({
           kind: NOTIFICATION_KINDS_SIDE.success,
           domain: DOMAINS.SIDE,
           text: intl.formatMessage(messages.methodDetailsStatusUpdated, {
-            methodName: formikValues.description,
+            methodName: formatLocalizedString(
+              {
+                name: formikValues.name,
+              },
+              {
+                key: 'name',
+                locale: dataLocale,
+                fallbackOrder: projectLanguages,
+                fallback: NO_VALUE_FALLBACK,
+              }
+            ),
             status: status === 'Active' ? 'activated' : 'deactivated',
           }),
         });
@@ -94,16 +118,21 @@ const MethodDetails = (props: TMethodDetailsProps) => {
   };
 
   const handleSubmitCallback = useCallback(handleSubmit, [
-    method,
+    method?.container,
+    method?.key,
     customObjectUpdater,
     showNotification,
     intl,
+    dataLocale,
+    projectLanguages,
   ]);
   const handleChangeCallback = useCallback(handleChange, [
     customObjectUpdater,
+    dataLocale,
     intl,
     method?.container,
     method?.key,
+    projectLanguages,
     showNotification,
   ]);
 
@@ -115,11 +144,42 @@ const MethodDetails = (props: TMethodDetailsProps) => {
       dataLocale={dataLocale}
     >
       {(formProps) => {
+        const methodName = formatLocalizedString(
+          {
+            name: formProps.values?.name,
+          },
+          {
+            key: 'name',
+            locale: dataLocale,
+            fallbackOrder: projectLanguages,
+            fallback: NO_VALUE_FALLBACK,
+          }
+        );
         return (
-          <CustomFormModalPage
-            title={formProps.values?.description ?? ''}
+          <TabularModalPage
+            title={methodName}
             isOpen
             onClose={() => props.onClose()}
+            tabControls={
+              <>
+                <TabHeader
+                  to={`${match.url}/general`}
+                  label="General setting"
+                  exactPathMatch={true}
+                />
+                <TabHeader
+                  isDisabled={true}
+                  to={`${match.url}/icon`}
+                  label="Icon"
+                />
+                <TabHeader
+                  isDisabled={false}
+                  to={`${match.url}/availability`}
+                  label="Availability"
+                  exactPathMatch={true}
+                />
+              </>
+            }
             formControls={
               <>
                 <SelectField
@@ -142,22 +202,25 @@ const MethodDetails = (props: TMethodDetailsProps) => {
                   ]}
                   horizontalConstraint={4}
                   controlShouldRenderValue={true}
-                  data-cy={'active-select'}
+                  data-testid={'status-select'}
                   isSearchable={false}
                 ></SelectField>
                 <CustomFormModalPage.FormSecondaryButton
                   label={CustomFormModalPage.Intl.revert}
                   onClick={formProps.handleReset}
                   isDisabled={!formProps.isDirty}
-                  dataAttributes={{ 'data-cy': 'revert-button' }}
+                  dataAttributes={{ 'data-testid': 'revert-button' }}
                 />
                 <CustomFormModalPage.FormPrimaryButton
                   label={CustomFormModalPage.Intl.save}
-                  onClick={() => formProps.submitForm()}
+                  onClick={(event) => {
+                    event.preventDefault();
+                    formProps.submitForm();
+                  }}
                   isDisabled={
                     formProps.isSubmitting || !formProps.isDirty || !canManage
                   }
-                  dataAttributes={{ 'data-cy': 'save-button' }}
+                  dataAttributes={{ 'data-testid': 'save-button' }}
                 />
               </>
             }
@@ -174,14 +237,20 @@ const MethodDetails = (props: TMethodDetailsProps) => {
                 </Text.Body>
               </ContentNotification>
             )}
-            {method && formProps.formElements}
-            {method && (
-              <ApplicationPageTitle
-                additionalParts={[formProps.values.description]}
-              />
-            )}
+            {method && <ApplicationPageTitle additionalParts={[methodName]} />}
             {method === null && <PageNotFound />}
-          </CustomFormModalPage>
+            <Switch>
+              <Route path={`${match.path}/general`}>
+                {method && formProps.formElements}
+              </Route>
+              <Route path={`${match.path}/icon`}>
+                <div>Icon</div>
+              </Route>
+              <Route path={`${match.path}/availability`}>
+                <div>Availability</div>
+              </Route>
+            </Switch>
+          </TabularModalPage>
         );
       }}
     </MethodDetailsForm>