+++ application/src/components/welcome/welcome.tsx
@@ -30,15 +30,22 @@ import { getErrorMessage } from '../../helpers';
 import { SuspendedRoute } from '@commercetools-frontend/application-shell';
 import MethodDetails from '../method-details';
 import { useIntl } from 'react-intl';
+import { formatLocalizedString } from '@commercetools-frontend/l10n';
+import { useApplicationContext } from '@commercetools-frontend/application-shell-connectors';
+import { NO_VALUE_FALLBACK } from '@commercetools-frontend/constants';
 
 const Welcome = () => {
   const intl = useIntl();
   const match = useRouteMatch();
   const { push } = useHistory();
+  const { dataLocale, projectLanguages } = useApplicationContext((context) => ({
+    dataLocale: context.dataLocale ?? '',
+    projectLanguages: context.project?.languages ?? [],
+  }));
   const columns = [
     {
-      key: 'description',
-      label: intl.formatMessage(messages.descriptionHeader),
+      key: 'name',
+      label: intl.formatMessage(messages.nameHeader),
     },
     {
       key: 'status',
@@ -58,7 +65,10 @@ const Welcome = () => {
       ),
     },
     { key: 'image', label: intl.formatMessage(messages.iconHeader) },
-    { key: 'order', label: intl.formatMessage(messages.displayOrderHeader) },
+    {
+      key: 'order',
+      label: intl.formatMessage(messages.displayOrderHeader),
+    },
   ];
   const customObjectUpdater = useCustomObjectDetailsUpdater();
   const { page, perPage } = usePaginationState();
@@ -78,7 +88,8 @@ const Welcome = () => {
   const [refresh, setRefresh] = useState<number>(0);
 
   const { fetchedData, fetchedDataLoading } = usePaymentMethodsFetcher(
-    extension?.destination?.url
+    extension?.destination?.url,
+    projectLanguages
   );
 
   const handleRefresh = useCallback(() => {
@@ -94,11 +105,15 @@ const Welcome = () => {
           );
 
           if (shouldCreate) {
-            await customObjectUpdater.execute({
-              container: OBJECT_CONTAINER_NAME,
-              key: method.id,
-              value: JSON.stringify(method),
-            });
+            await customObjectUpdater
+              .execute({
+                container: OBJECT_CONTAINER_NAME,
+                key: method.id,
+                value: JSON.stringify(method),
+              })
+              .catch((error) => {
+                console.error(`Error creating custom object: ${error}`);
+              });
             return method;
           } else {
             return customObjectsPaginatedResult?.results.find(
@@ -163,8 +178,20 @@ const Welcome = () => {
                 ) : (
                   <CheckInactiveIcon color="neutral60"></CheckInactiveIcon>
                 );
-              case 'description':
-                return item.description;
+              case 'name':
+                return item.name
+                  ? formatLocalizedString(
+                      {
+                        name: item.name,
+                      },
+                      {
+                        key: 'name',
+                        locale: dataLocale,
+                        fallbackOrder: projectLanguages,
+                        fallback: NO_VALUE_FALLBACK,
+                      }
+                    )
+                  : item.description;
               case 'image':
                 return (
                   <IconButton
@@ -182,7 +209,7 @@ const Welcome = () => {
                   ></IconButton>
                 );
               case 'order':
-                return item.displayOrder;
+                return item.displayOrder ?? '-';
               default:
                 return null;
             }
@@ -196,7 +223,7 @@ const Welcome = () => {
                 customObjectsPaginatedResult?.results.filter(
                   (obj) => obj.key === row.id
                 )?.[0]?.id
-              }`
+              }/general`
             );
           }}
         />