+++ application/src/types.ts
@@ -23,8 +23,9 @@ export type TMethodObjectFormValues = {
 
 export type TMethodObjectValueFormValues = {
   id: string;
-  description: string;
+  name: Record<string, string>;
+  description: Record<string, string>;
   imageUrl: string;
   status: string;
-  displayOrder?: number;
+  displayOrder: number;
 };